Add storage functionality
To store and update datums correctly, a storage class has been added.
This should help keep the stored datums up to date and provide an
unified API to the user. Additionally it makes sure that there are no
gaps in the time series, and if so it automatically created a new entry,
so nothing is lost. To make the user aware of such gabs, the creation of
a new storage file is logged as information.