Add update script for datums
To allow the server to keep its data warehouse up to date I've added a
simple update script. While testing it on the real Kraken API some bugs
became apparent, and additional tests and features have been added such
as actually storing the csv value after querying it from the source.

Also there seem to be still some problems regarding the statistical
makeup of some of the real world data which needs to be further
investigates.