Add trades cache to kraken source
Querying the Kraken API can take quite long as there is a 6 seconds
limit on trades queries and each query just returns 1000 elements. To
avoid having to download time and again data from the Kraken server we
cache trades in a simple .gz file.

The cache currently is quite slow, but also small. We might be able to
improve performance with an index.