Prepare kraken adapter for cache refactoring
We moved the response validation logic to the kraken source as the class
that actually does the remote query, should also be responsible for
validating it. Further the adapter has been simplified to just take a
list of trades, and processing the raw response to trades has been moved
to the source class, as this simplifies the cache.