Improve cache performance
As trades data can become quite large over the years, the cache is zip
compressed when stored. To still allow for fast queries to header also
stores the last timestamp rounded up to the next integer. This way it is
possible to skip chunks of trade data until we actually reach the lower
bound of the query.